requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
abseil-cpp=20210324.2=h2531618_0
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py37h7f8727e_2
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=3.0.0=py37h6b21186_4
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bcj-cffi=0.5.1=py37h295c915_0
black=19.10b0=py_0
blas=1.0=mkl
blinker=1.4=py_1
boost-cpp=1.73.0=h27cfd23_11
bottleneck=1.3.4=py37hce1f21e_0
brotli=1.0.9=he6710b0_2
brotli-python=1.0.9=py37heb0550a_2
brotlicffi=1.0.9.2=py37h295c915_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py37h89c1867_2
cffi=1.15.0=py37hd667e15_1
chardet=4.0.0=py37h89c1867_2
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37h89c1867_0
colorama=0.4.4=pyh9f0ad1d_0
conllu=4.4.1=pyhd3eb1b0_0
cryptography=36.0.0=py37h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=1.12.1=pyhd3eb1b0_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.4=pyhd3eb1b0_0
double-conversion=3.1.5=he6710b0_1
entrypoints=0.4=pyhd8ed1ab_0
et_xmlfile=1.1.0=py37h06a4308_0
ffmpeg=4.3=hf484d3e_0
filelock=3.6.0=pyhd8ed1ab_0
fonttools=4.32.0=pypi_0
freetype=2.11.0=h70c0345_0
fsspec=2022.2.0=pyhd8ed1ab_0
future=0.18.2=py37h89c1867_4
fuzzywuzzy=0.18.0=pypi_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h7b6447c_0
glog=0.5.0=h2531618_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpc-cpp=1.39.0=hae934f6_5
grpcio=1.42.0=py37hce63b2e_0
huggingface-hub=0.2.1=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py37h89c1867_0
importlib_metadata=4.11.3=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.10.0=py37h25bab4e_0
ipython=7.32.0=py37h89c1867_0
jedi=0.18.1=py37h89c1867_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h7f8727e_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py37h89c1867_0
kiwisolver=1.4.2=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libboost=1.73.0=h3ff78a5_11
libcurl=7.80.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h36c2ea0_1
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libthrift=0.14.2=hcc01f38_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxml2=2.9.12=h03d6c58_0
libxslt=1.1.34=hc22bd24_0
lxml=4.8.0=py37h1f438cf_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=5.2.0=py37h7f8727e_2
multiprocess=0.70.12.2=py37h7f8727e_0
multivolumefile=0.2.3=pyhd3eb1b0_0
mypy_extensions=0.4.3=py37h06a4308_1
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
nltk=3.7=pyhd3eb1b0_0
numexpr=2.8.1=py37h6abb31d_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
oauthlib=3.2.0=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1n=h7f8727e_0
orc=1.6.9=ha97a36c_3
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=py37h8c16a72_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=py37h22f2fdc_0
pip=21.2.2=py37h06a4308_0
prompt-toolkit=3.0.27=pyha770c72_0
protobuf=3.17.2=py37h295c915_0
psutil=5.8.0=py37h27cfd23_1
ptyprocess=0.7.0=pyhd3deb0d_0
py-rouge=1.1=pypi_0
py7zr=0.16.1=pyhd3eb1b0_1
pyarrow=3.0.0=py37he0739d4_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.10.1=py37h27cfd23_1
pydeprecate=0.3.2=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyppmd=0.16.1=py37h295c915_0
pysocks=1.7.1=py37_1
python=3.7.10=h12debd9_4
python-dateutil=2.8.2=pyhd8ed1ab_0
python-levenshtein=0.12.2=pypi_0
python-xxhash=2.0.2=py37h7f8727e_0
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.5.10=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=5.4.1=py37h5e8e339_0
pyzmq=19.0.2=py37hac76be4_2
pyzstd=0.14.4=py37h7f8727e_3
re2=2020.11.01=h2531618_1
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py37h7f8727e_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rouge=1.0.1=pypi_0
rouge-score=0.0.4=pypi_0
rsa=4.8=pyhd8ed1ab_0
sacremoses=0.0.49=pyhd8ed1ab_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.8=he6710b0_0
sqlite=3.38.0=hc218d9a_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py37h7f0c10b_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
texttable=1.6.4=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.11.6=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.11.0=py37_cu102
torchinfo=1.6.5=pyhd8ed1ab_0
torchmetrics=0.7.3=pyhd8ed1ab_0
torchvision=0.12.0=py37_cu102
tornado=6.1=py37h5e8e339_1
tqdm=4.62.1=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.17.0=pypi_0
typed-ast=1.4.3=py37h7f8727e_1
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
uriparser=0.9.3=he6710b0_1
urllib3=1.26.8=pyhd3eb1b0_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
xxhash=0.8.0=h7f8727e_3
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py37h27cfd23_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
